fix(validator): cache `arrayBuffer` to use after validation
