+++ src/validator/validator.ts
@@ -1,6 +1,7 @@
 import type { Context } from '../context'
 import { getCookie } from '../helper/cookie'
 import type { Env, ValidationTargets, MiddlewareHandler } from '../types'
+import type { BodyData } from '../utils/body'
 
 type ValidationTargetKeysWithBody = 'form' | 'json'
 type ValidationTargetByMethod<M> = M extends 'get' | 'head' // GET and HEAD request must not have a body content.
@@ -54,6 +55,7 @@ export const validator = <
            */
           const arrayBuffer = c.req.bodyCache.arrayBuffer ?? (await c.req.raw.arrayBuffer())
           value = await new Response(arrayBuffer).json()
+          c.req.bodyCache.json = value
           c.req.bodyCache.arrayBuffer = arrayBuffer
         } catch {
           console.error('Error: Malformed JSON in request body')
@@ -67,7 +69,20 @@ export const validator = <
         }
         break
       case 'form':
-        value = await c.req.parseBody()
+        await (async () => {
+          const contentType = c.req.header('Content-Type')
+          if (contentType) {
+            const arrayBuffer = c.req.bodyCache.arrayBuffer ?? (await c.req.raw.arrayBuffer())
+            const formData = arrayBufferToFormData(arrayBuffer, contentType)
+            const form: BodyData = {}
+            formData.forEach((value, key) => {
+              form[key] = value
+            })
+            value = form
+            c.req.bodyCache.formData = formData
+            c.req.bodyCache.arrayBuffer = arrayBuffer
+          }
+        })()
         break
       case 'query':
         value = Object.fromEntries(
@@ -101,4 +116,35 @@ export const validator = <
 
     await next()
   }
+}
+
+const _decodeURIComponent = decodeURIComponent
+
+const arrayBufferToFormData = (arrayBuffer: ArrayBuffer, contentType: string) => {
+  const decoder = new TextDecoder('utf-8')
+  const content = decoder.decode(arrayBuffer)
+  const formData = new FormData()
+
+  const boundaryMatch = contentType.match(/boundary=(.+)/)
+  const boundary = boundaryMatch ? boundaryMatch[1] : ''
+
+  if (contentType.startsWith('multipart/form-data') && boundary) {
+    const parts = content.split('--' + boundary).slice(1, -1)
+    for (const part of parts) {
+      const [header, body] = part.split('\r\n\r\n')
+      const nameMatch = header.match(/name="([^"]+)"/)
+      if (nameMatch) {
+        const name = nameMatch[1]
+        formData.append(name, body.trim())
+      }
+    }
+  } else if (contentType.startsWith('application/x-www-form-urlencoded')) {
+    const pairs = content.split('&')
+    for (const pair of pairs) {
+      const [key, value] = pair.split('=')
+      formData.append(_decodeURIComponent(key), _decodeURIComponent(value))
+    }
+  }
+
+  return formData
 }
